Draft: doc: (fixes #1256) update to Doxygen v1.15

Closes #1256

See merge request nsnam/ns-3-dev!2563

+++ src/traffic-control/model/cobalt-queue-disc.cc
@@ -12,6 +12,11 @@
  *                    Shefali Gupta <shefaligups11@gmail.com>
  *                    Mohit P. Tahiliani <tahiliani@nitk.edu.in>
  */
+/**
+ * @file
+ * @ingroup traffic-control
+ * Class ns3::CobaltQueueDisc implementation.
+ */
 
 #include "cobalt-queue-disc.h"
 
@@ -230,7 +235,7 @@ CobaltQueueDisc::CacheInit()
     m_recInvSqrt = ~0U;
     m_recInvSqrtCache[0] = m_recInvSqrt;
 
-    for (m_count = 1; m_count < (uint32_t)(REC_INV_SQRT_CACHE); m_count++)
+    for (m_count = 1; m_count < REC_INV_SQRT_CACHE; m_count++)
     {
         NewtonStep();
         NewtonStep();
@@ -243,7 +248,7 @@ CobaltQueueDisc::CacheInit()
 void
 CobaltQueueDisc::InvSqrt()
 {
-    if (m_count < (uint32_t)REC_INV_SQRT_CACHE)
+    if (m_count < REC_INV_SQRT_CACHE)
     {
         m_recInvSqrt = m_recInvSqrtCache[m_count];
     }
